rename: PlayerSetting 클래스 패키지 이동
오류로 인하여 원인을 찾는중
